refactor: useTripItemExpenseInput 커스텀 훅 hooks/trip 폴더로 이동
